Add proper check for empty content before displaying UI
There are cases, when providers will return an content table with an
empty string for display. This caused it to fall through the is empty
check, causing hovercraft to crash, as the opened popup would have
a width of `0`.

To fix this, there is now a more elaborate `_is_empty_content` check, as
well as some tests to ensure it is working as expected.

Fixes #3