Add compat mode to work properly with render-markdown.nvim plugin
When hovercraft.nvim is used with the render-markdown.nvim plugin, the
floating window is one line short. To work around this annoying issue,
a new compat mode has been introduced. If the
`window.render_markdown_compat_mode` setting is set, then a dummy empty
line will be appended to the end of the floating window. It is a bit of
a hacky workaround, but the simplest solution for now.

Fixes #10