Fix nested open of hovercraft.nvim popup
Hovercraft now handles the case it is opened from within a hovercraft
popup. When the case happens, the target of the autogroup responsible
for handling closing of the popup is then set to the underlying current
buffer. Overall the solution is a bit of a hack, but good enough for
now. In the future more brain power should be put into how to properly
handle such cases (e.g. disallow nested popups).

Partially fixes #11